@@ -22,16 +22,16 @@ repositories {
 }
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     // buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.sunoray.tentacle"
         minSdkVersion 16
-        targetSdkVersion 26
-        versionCode 40
-        versionName "2.0.7"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 52
+        versionName "2.0.13"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         //jackOptions.enabled = true
     }
 
@@ -69,11 +69,11 @@ dependencies {
     //androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
     //    exclude group: 'com.android.support', module: 'support-annotations'
     //})
-    implementation 'com.android.support:appcompat-v7:26.+'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     //testCompile 'junit:junit:4.12'
-    implementation 'com.android.support:design:26.+'
-    implementation 'com.android.support:support-v4:26.+'
-    implementation 'com.android.support:support-v13:26.+'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    //implementation 'com.android.support:support-v13:28.+'
     implementation 'com.google.code.gson:gson:2.7'
 
     // log dependency
