@@ -4,8 +4,11 @@ import android.Manifest;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.RequiresApi;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.crashlytics.android.Crashlytics;
 import com.tbruyelle.rxpermissions.RxPermissions;
@@ -26,7 +29,10 @@ public class SplashScreen extends AppCompatActivity {
             public void run() {
                 try {
                     sleep(1 * 1000);
-                    CheckPermissions();
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+                        CheckPermissionsWithAnswerCall();
+                    else
+                        CheckPermissions();
                 } catch (Exception e) {
                 }
             }
@@ -76,6 +82,37 @@ public class SplashScreen extends AppCompatActivity {
                         Manifest.permission.GET_ACCOUNTS,
                         Manifest.permission.ACCESS_FINE_LOCATION,
                         Manifest.permission.ACCESS_COARSE_LOCATION
+                )
+                .subscribe(new Action1<Boolean>() {
+                    @Override
+                    public void call(Boolean aBoolean) {
+                        initialize(aBoolean);
+                    }
+                }, new Action1<Throwable>() {
+                    @Override
+                    public void call(Throwable throwable) {
+
+                    }
+                });
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.O)
+    void CheckPermissionsWithAnswerCall() {
+        /*android.Manifest.permission.WRITE_EXTERNAL_STORAGE,*/
+        RxPermissions.getInstance(SplashScreen.this)
+                .request(
+                        Manifest.permission.READ_PHONE_STATE,
+                        Manifest.permission.CALL_PHONE,
+                        Manifest.permission.READ_CALL_LOG,
+                        Manifest.permission.WRITE_CALL_LOG,
+                        Manifest.permission.READ_EXTERNAL_STORAGE,
+                        Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                        Manifest.permission.RECORD_AUDIO,
+                        Manifest.permission.READ_CONTACTS,
+                        Manifest.permission.GET_ACCOUNTS,
+                        Manifest.permission.ACCESS_FINE_LOCATION,
+                        Manifest.permission.ACCESS_COARSE_LOCATION,
+                        Manifest.permission.ANSWER_PHONE_CALLS
                 )
                 .subscribe(new Action1<Boolean>() {
                     @Override
