@@ -9,6 +9,7 @@ import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.AlertDialog;
+import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -22,10 +23,11 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.StrictMode;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.v13.app.ActivityCompat;
-import android.support.v4.content.WakefulBroadcastReceiver;
-import android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener;
+
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.legacy.content.WakefulBroadcastReceiver;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout.OnRefreshListener;
 import android.view.View;
 import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
@@ -284,7 +286,7 @@ public class ViewActivity extends Activity {
                             }
                     ).show();
         } catch (Exception e) {
-            log.debug("Exception in ImageChooserDialog ",e);
+            log.debug("Exception in ImageChooserDialog ", e);
         }
 
     }
@@ -450,9 +452,34 @@ public class ViewActivity extends Activity {
                 view.loadUrl(url);
                 return true;
             } else {
-                Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
-                startActivity(browserIntent);
+                try {
+                    Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+                    if (browserIntent.resolveActivity(getPackageManager()) != null)
+                        startActivity(browserIntent);
+                    else
+                        activityNotFoundAlert(ViewActivity.this);
+                } catch (ActivityNotFoundException e) {
+                    log.info("Exception (startActivity): ", e);
+                    activityNotFoundAlert(ViewActivity.this);
+                } catch (Exception e) {
+                    log.info("Exception (startActivity): ", e);
+                }
                 return true;
+            }
+        }
+
+        private void activityNotFoundAlert(Context context) {
+            try {
+                new AlertDialog.Builder(context, Util.getAlertTheame()).setTitle("Telephony Service Not Found")
+                        .setMessage("The application configured in custom telephony service for making calls is not found. Make sure application is installed before making the call.")
+                        .setCancelable(false)
+                        .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int id) {
+                                dialog.cancel();
+                            }
+                        }).create().show();
+            } catch (Exception e) {
+                log.info("Exception in activityNotFoundAlert", e);
             }
         }
 
