@@ -6,13 +6,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.sunoray.tentacle.common.AppProperties;
+import com.sunoray.tentacle.common.Util;
 
 import android.Manifest;
 import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.pm.PackageManager;
 import android.os.Build;
-import android.support.v4.app.ActivityCompat;
+import androidx.core.app.ActivityCompat;
 import android.telecom.TelecomManager;
 import android.telephony.TelephonyManager;
 
@@ -47,12 +48,7 @@ public class CallHelper {
 	public static void endCall(Context context) {
 		try {
 			if(isCallInProgress(context)) {
-				TelephonyManager tm = (TelephonyManager) context.getSystemService(Context.TELEPHONY_SERVICE);
-				Method m1 = tm.getClass().getDeclaredMethod("getITelephony");
-				m1.setAccessible(true);
-				Object iTelephony = m1.invoke(tm);
-				Method m2 = iTelephony.getClass().getDeclaredMethod("endCall");
-				m2.invoke(iTelephony);
+				Util.endCall(context);
 				log.info("Call Disconnected");	
 			}
 		} catch (Exception e) {
