@@ -13,7 +13,7 @@ import android.app.Service;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.IBinder;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 
 import com.sunoray.tentacle.R;
 import com.sunoray.tentacle.StartupActivity;
@@ -81,7 +81,7 @@ public class BackGroundService extends Service {
 				.setContentText(filename)
 				.setContentIntent(pendingIntent)
 				.setSmallIcon(R.drawable.ic_launcher)
-				.setTicker("Listening calls")
+				.setTicker("Monitoring calls")
 				.setPriority(NotificationCompat.PRIORITY_MIN)
 				.build();
 
